Green, not yellow!
I added primaries. This one was fun. I got to tell the computer to give me the eigenvalue of Washington state (it's 3, if you were wondering). I also realised that my disqualification system from the primaries was actually converting candidates into extremists instead of disqualifying them, so I fixed that.